@@ -2,12 +2,12 @@ name: examples
 description: dart.dev example code.
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   args: ^2.3.0
   examples_util: { path: ../util }
 
 dev_dependencies:
-  lints: ^1.0.0
+  lints: ^2.0.0
   test: ^1.19.2
