@@ -35,6 +35,14 @@ void miscDeclAnalyzedButNotTested() {
     // #docregion use-named-parameters
     enableFlags(bold: true, hidden: false);
     // #enddocregion use-named-parameters
+
+    void repeat(Function func, {int times = 1}) {}
+
+    // #docregion named-arguments-anywhere
+    repeat(times: 2, () {
+      /*...*/
+    });
+    // #enddocregion named-arguments-anywhere
   }
 
   {
@@ -84,7 +92,6 @@ abstract class Widget {
 
 class Scrollbar extends Widget {
   // #docregion required-named-parameters
-  const Scrollbar({Key? key, required Widget child})
-      // #enddocregion required-named-parameters
-      : super(key: key);
+  const Scrollbar({super.key, required Widget child});
+  // #enddocregion required-named-parameters
 }
