@@ -16,7 +16,8 @@ void main() {
     querySelector('#confirm') // Get an object.
       ?..text = 'Confirm' // Use its members.
       ..classes.add('important')
-      ..onClick.listen((e) => window.alert('Confirmed!'));
+      ..onClick.listen((e) => window.alert('Confirmed!'))
+      ..scrollIntoView();
     // #enddocregion cascade-operator
 
     expect(document.querySelector('#confirm')?.text, 'Confirm');
@@ -31,6 +32,7 @@ void main() {
     button?.text = 'Confirm';
     button?.classes.add('important');
     button?.onClick.listen((e) => window.alert('Confirmed!'));
+    button?.scrollIntoView();
     // #enddocregion cascade-operator-example-expanded
 
     expect(document.querySelector('#confirm')?.text, 'Confirm');
